+++ server/src/test/java/com/vaadin/tests/data/converter/StringToFloatConverterTest.java
@@ -29,7 +29,7 @@ public class StringToFloatConverterTest extends AbstractStringConverterTest {
 
     @Test
     public void testValueConversion() {
-        assertValue(Float.valueOf(10),
+        assertValue(10F,
                 getConverter().convertToModel("10", new ValueContext()));
     }
